Move Tasks to MissionTasks and add GitStore as an abstraction of AssetGroups
