+++ app/modules/asset_groups/models.py
@@ -7,60 +7,25 @@ import copy
 import enum
 from flask import current_app
 from flask_login import current_user  # NOQA
-import requests.exceptions
-import utool as ut
-import git
 from datetime import datetime  # NOQA
-from git import Git as BaseGit, Repo as BaseRepo
-
-from app.extensions.gitlab import GitlabInitializationError
-from app.extensions import db, HoustonModel, parallel
+from app.extensions import db, HoustonModel
 import app.extensions.logging as AuditLog  # NOQA
+from app.extensions.git_store import GitStore
 from app.modules.annotations.models import Annotation
 from app.modules.assets.models import Asset
 from app.modules.encounters.models import Encounter
 from app.modules.sightings.models import Sighting, SightingStage
 from app.modules.users.models import User
 from app.utils import HoustonException
-from app.version import version
 
 import logging
-import tqdm
 import uuid
 import json
-import os
-import pathlib
-import shutil
 from urllib.parse import urljoin
 
 from .metadata import AssetGroupMetadata
 
 log = logging.getLogger(__name__)  # pylint: disable=invalid-name
-
-
-def compute_xxhash64_digest_filepath(filepath):
-    try:
-        import xxhash
-        import os
-
-        assert os.path.exists(filepath)
-
-        with open(filepath, 'rb') as file_:
-            digest = xxhash.xxh64_hexdigest(file_.read())
-    except Exception:
-        digest = None
-    return digest
-
-
-class AssetGroupMajorType(str, enum.Enum):
-    filesystem = 'filesystem'
-    archive = 'archive'
-    service = 'service'
-    test = 'test'
-
-    unknown = 'unknown'
-    error = 'error'
-    reject = 'reject'
 
 
 class AssetGroupSightingStage(str, enum.Enum):
@@ -69,35 +34,6 @@ class AssetGroupSightingStage(str, enum.Enum):
     curation = 'curation'
     processed = 'processed'
     failed = 'failed'
-
-
-class _Git(BaseGit):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Assign the SSH command to include the id key file for authentication
-        self.update_environment(GIT_SSH_COMMAND=current_app.config['GIT_SSH_COMMAND'])
-
-
-class Repo(BaseRepo):
-    GitCommandWrapperType = _Git
-
-    # This class is copied from the original with one minor adjustment
-    # to use the class' `GitCommandWrapperType` definition as opposed
-    # to the direct reference to the `Git` class.
-    # Fixed in https://github.com/gitpython-developers/GitPython/pull/1322
-    @classmethod
-    def clone_from(
-        cls, url, to_path, progress=None, env=None, multi_options=None, **kwargs
-    ):
-        # git = Git(os.getcwd())
-        git = cls.GitCommandWrapperType(os.getcwd())
-        if env is not None:
-            git.update_environment(**env)
-        from git.db import GitCmdObjectDB
-
-        return cls._clone(
-            git, url, to_path, GitCmdObjectDB, progress, multi_options, **kwargs
-        )
 
 
 # AssetGroup can have many sightings, so needs a table
@@ -822,76 +758,16 @@ class AssetGroupSighting(db.Model, HoustonModel):
         return self.config.get('idConfigs', [])
 
 
-class AssetGroup(db.Model, HoustonModel):
+class AssetGroup(GitStore):
     """
     AssetGroup database model.
-
-    AssetGroup Structure:
-        _db/asset_groups/<asset_group GUID>/
-            - .git/
-            - _asset_group/
-            - - <user's uploaded data>
-            - _assets/
-            - - <symlinks into _asset_group/ folder> with name <asset GUID >.ext --> ../_asset_group/path/to/asset/original_name.ext
-            - metadata.json
     """
 
-    __mapper_args__ = {
-        'confirm_deleted_rows': False,
-    }
+    GIT_STORE_NAME = 'asset_groups'
 
-    guid = db.Column(
-        db.GUID, default=uuid.uuid4, primary_key=True
-    )  # pylint: disable=invalid-name
+    GIT_STORE_DATABASE_PATH_CONFIG_NAME = 'ASSET_GROUP_DATABASE_PATH'
 
-    major_type = db.Column(
-        db.Enum(AssetGroupMajorType),
-        default=AssetGroupMajorType.unknown,
-        index=True,
-        nullable=False,
-    )
-
-    commit = db.Column(db.String(length=40), nullable=True, unique=True)
-    commit_mime_whitelist_guid = db.Column(db.GUID, index=True, nullable=True)
-    commit_houston_api_version = db.Column(db.String, index=True, nullable=True)
-
-    description = db.Column(db.String(length=255), nullable=True)
-
-    config = db.Column(db.JSON, default=lambda: {}, nullable=False)
-
-    owner_guid = db.Column(
-        db.GUID, db.ForeignKey('user.guid'), index=True, nullable=False
-    )
-    # owner = db.relationship(
-    #     'User', backref='asset_groups', foreign_keys=[owner_guid]
-    # )
-    owner = db.relationship(
-        'User',
-        backref=db.backref(
-            'asset_groups',
-            primaryjoin='User.guid == AssetGroup.owner_guid',
-            order_by='AssetGroup.guid',
-        ),
-        foreign_keys=[owner_guid],
-    )
-
-    submitter_guid = db.Column(
-        db.GUID, db.ForeignKey('user.guid'), index=True, nullable=True
-    )
-    submitter = db.relationship(
-        'User',
-        back_populates='submitted_asset_groups',
-        foreign_keys=[submitter_guid],
-    )
-    submitter = db.relationship(
-        'User',
-        backref=db.backref(
-            'submitted_asset_groups',
-            primaryjoin='User.guid == AssetGroup.submitter_guid',
-            order_by='AssetGroup.guid',
-        ),
-        foreign_keys=[submitter_guid],
-    )
+    guid = db.Column(db.GUID, db.ForeignKey('git_store.guid'), primary_key=True)
 
     asset_group_sightings = db.relationship(
         'AssetGroupSighting',
@@ -900,662 +776,33 @@ class AssetGroup(db.Model, HoustonModel):
         cascade='all, delete',
     )
 
-    # assets = db.relationship(
-    #     'Asset', back_populates='asset_group', order_by='Asset.guid'
-    # )
+    __mapper_args__ = {
+        'polymorphic_identity': 'asset_group',
+    }
 
-    def __repr__(self):
-        return (
-            '<{class_name}('
-            'guid={self.guid}, '
-            ')>'.format(class_name=self.__class__.__name__, self=self)
-        )
+    @classmethod
+    def ensure_remote_delay(cls, asset_group):
+        from .tasks import ensure_remote
 
-    @property
-    def bulk_upload(self):
-        return isinstance(self.config, dict) and self.config.get('uploadType') == 'bulk'
+        ensure_remote.delay(str(asset_group.guid))
 
-    @property
-    def anonymous(self):
-        return self.owner is User.get_public_user()
+    def git_push_delay(self):
+        from .tasks import git_push
 
-    @property
-    def mime_type_whitelist(self):
-        if getattr(self, '_mime_type_whitelist', None) is None:
-            asset_mime_type_whitelist = current_app.config.get(
-                'ASSET_MIME_TYPE_WHITELIST', []
-            )
-            asset_mime_type_whitelist = sorted(list(map(str, asset_mime_type_whitelist)))
+        git_push.delay(str(self.guid))
 
-            self._mime_type_whitelist = set(asset_mime_type_whitelist)
-        return self._mime_type_whitelist
+    def delete_remote_delay(self):
+        from .tasks import delete_remote
 
-    @property
-    def mime_type_whitelist_guid(self):
-        if getattr(self, '_mime_type_whitelist_guid', None) is None:
-            self._mime_type_whitelist_guid = ut.hashable_to_uuid(
-                sorted(list(self.mime_type_whitelist))
-            )
-            # Write mime.whitelist.<mime-type-whitelist-guid>.json
-            mime_type_whitelist_mapping_filepath = os.path.join(
-                current_app.config.get('PROJECT_DATABASE_PATH'),
-                'mime.whitelist.%s.json' % (self._mime_type_whitelist_guid,),
-            )
-            if not os.path.exists(mime_type_whitelist_mapping_filepath):
-                log.debug(
-                    'Creating new MIME whitelist manifest: %r'
-                    % (mime_type_whitelist_mapping_filepath,)
-                )
-                with open(mime_type_whitelist_mapping_filepath, 'w') as mime_type_file:
-                    mime_type_whitelist_dict = {
-                        str(self._mime_type_whitelist_guid): sorted(
-                            list(self.mime_type_whitelist)
-                        ),
-                    }
-                    mime_type_file.write(json.dumps(mime_type_whitelist_dict))
-        return self._mime_type_whitelist_guid
+        delete_remote.delay(str(self.guid))
 
-    def get_config_field(self, field):
-        return self.config.get(field) if isinstance(self.config, dict) else None
-
-    def _ensure_repository_files(self):
-        group_path = self.get_absolute_path()
-
-        # AssetGroup Repo Structure:
-        #     _db/assetGroup/<asset_group GUID>/
-        #         - .git/
-        #         - _asset_group/
-        #         - - <user's uploaded data>
-        #         - _assets/
-        #         - - <symlinks into _asset_group/ folder> with name <asset GUID >.ext --> ../_asset_group/path/to/asset/original_name.ext
-        #         - metadata.json
-
-        if not os.path.exists(group_path):
-            # Initialize local repo
-            log.info('Creating asset_groups structure: %r' % (group_path,))
-            os.mkdir(group_path)
-
-        # Create the repo
-        git_path = os.path.join(group_path, '.git')
-        if not os.path.exists(git_path):
-            repo = Repo.init(group_path)
-            assert len(repo.remotes) == 0
-            git_remote_public_name = current_app.config.get('GIT_PUBLIC_NAME', None)
-            git_remote_email = current_app.config.get('GIT_EMAIL', None)
-            assert None not in [git_remote_public_name, git_remote_email]
-            repo.git.config('user.name', git_remote_public_name)
-            repo.git.config('user.email', git_remote_email)
-        else:
-            repo = Repo(group_path)
-
-        asset_group_path = os.path.join(group_path, '_asset_group')
-        if not os.path.exists(asset_group_path):
-            os.mkdir(asset_group_path)
-        pathlib.Path(os.path.join(asset_group_path, '.touch')).touch()
-
-        assets_path = os.path.join(group_path, '_assets')
-        if not os.path.exists(assets_path):
-            os.mkdir(assets_path)
-        pathlib.Path(os.path.join(assets_path, '.touch')).touch()
-
-        metadata_path = os.path.join(group_path, 'metadata.json')
-        if not os.path.exists(metadata_path):
-            with open(metadata_path, 'w') as metatdata_file:
-                json.dump({}, metatdata_file)
-
-        return repo
-
-    def git_write_upload_file(self, upload_file):
-        repo = self.ensure_repository()
-        file_repo_path = os.path.join(
-            repo.working_tree_dir, '_asset_group', upload_file.filename
-        )
-        upload_file.save(file_repo_path)
-        log.info('Wrote file upload and added to local repo: %r' % (file_repo_path,))
-
-    def git_copy_path(self, path):
-        absolute_path = os.path.abspath(os.path.expanduser(path))
-        if not os.path.exists(path):
-            raise IOError('The path %r does not exist.' % (absolute_path,))
-
-        repo = self.ensure_repository()
-        repo_path = os.path.join(repo.working_tree_dir, '_asset_group')
-
-        absolute_path = absolute_path.rstrip('/')
-        repo_path = repo_path.rstrip('/')
-        absolute_path = '%s/' % (absolute_path,)
-        repo_path = '%s/' % (repo_path,)
-
-        if os.path.exists(repo_path):
-            shutil.rmtree(repo_path)
-
-        shutil.copytree(absolute_path, repo_path)
-
-    def git_copy_file_add(self, filepath):
-        absolute_filepath = os.path.abspath(os.path.expanduser(filepath))
-        if not os.path.exists(absolute_filepath):
-            raise IOError('The filepath %r does not exist.' % (absolute_filepath,))
-
-        repo = self.ensure_repository()
-        repo_path = os.path.join(repo.working_tree_dir, '_asset_group')
-        _, filename = os.path.split(absolute_filepath)
-        repo_filepath = os.path.join(repo_path, filename)
-
-        shutil.copyfile(absolute_filepath, repo_filepath)
-
-        return repo_filepath
-
-    def git_commit(self, message, realize=True, update=True, **kwargs):
-        repo = self.ensure_repository()
-
-        if realize:
-            self.realize_asset_group()
-
-        if update:
-            self.update_asset_symlinks(**kwargs)
-
-        asset_group_path = self.get_absolute_path()
-        asset_group_metadata_path = os.path.join(asset_group_path, 'metadata.json')
-
-        assert os.path.exists(asset_group_metadata_path)
-        with open(asset_group_metadata_path, 'r') as asset_group_metadata_file:
-            asset_group_metadata = json.load(asset_group_metadata_file)
-
-        asset_group_metadata['commit_mime_whitelist_guid'] = str(
-            self.mime_type_whitelist_guid
-        )
-        asset_group_metadata['commit_houston_api_version'] = str(version)
-
-        if 'frontend_sightings_data' not in asset_group_metadata and self.config:
+    def git_commit_metadata_hook(self, local_store_metadata):
+        if 'frontend_sightings_data' not in local_store_metadata and self.config:
             metadata_request = self.config
             metadata_request['sightings'] = []
             for sighting in self.asset_group_sightings:
                 metadata_request['sightings'].append(sighting.config)
-
-            asset_group_metadata['frontend_sightings_data'] = metadata_request
-
-        with open(asset_group_metadata_path, 'w') as asset_group_metadata_file:
-            json.dump(asset_group_metadata, asset_group_metadata_file)
-
-        # repo.index.add('.gitignore')
-        repo.index.add('_assets/')
-        repo.index.add('_asset_group/')
-        repo.index.add('metadata.json')
-
-        commit = repo.index.commit(message)
-
-        self.update_metadata_from_commit(commit)
-
-    def git_pull(self):
-        repo = self.get_repository()
-        assert repo is not None
-
-        log.info('Pulling from remote repository')
-        try:
-            repo.git.pull(repo.remotes.origin, repo.head.ref)
-        except git.exc.GitCommandError as e:
-            log.info(f'git pull failed for {self.guid}: {str(e)}')
-        else:
-            log.info('...pulled')
-
-        self.update_metadata_from_repo(repo)
-
-        return repo
-
-    def git_clone(self, project, **kwargs):
-        repo = self.get_repository()
-        assert repo is None
-
-        asset_group_abspath = self.get_absolute_path()
-        remote_url = project.ssh_url_to_repo
-
-        args = (
-            remote_url,
-            asset_group_abspath,
-        )
-        log.info('Cloning remote asset_group:\n\tremote: %r\n\tlocal:  %r' % args)
-        repo = Repo.clone_from(remote_url, asset_group_abspath)
-        log.info('...cloned')
-
-        repo = self.get_repository()
-        assert repo is not None
-
-        self.update_metadata_from_project(project)
-        self.update_metadata_from_repo(repo)
-
-        # Traverse the repo and create Asset objects in database
-        self.update_asset_symlinks(**kwargs)
-
-        return repo
-
-    @classmethod
-    def ensure_asset_group(cls, asset_group_uuid, owner=None):
-        from .tasks import ensure_remote
-
-        asset_group = AssetGroup.query.get(asset_group_uuid)
-        if asset_group is None:
-            from app.extensions import db
-
-            if not AssetGroup.is_on_remote(asset_group_uuid):
-                return None
-
-            if owner is None:
-                owner = current_user
-
-            asset_group = AssetGroup(
-                guid=asset_group_uuid,
-                owner_guid=owner.guid,
-            )
-
-            with db.session.begin():
-                db.session.add(asset_group)
-            db.session.refresh(asset_group)
-
-        # Make sure that the repo for this asset group exists
-        asset_group.ensure_repository()
-        # Create gitlab project in the background (we won't wait for its
-        # completion here)
-        ensure_remote.delay(str(asset_group.guid))
-
-        return asset_group
-
-    @classmethod
-    def create_from_metadata(cls, metadata):
-        if metadata.owner is not None and not metadata.owner.is_anonymous:
-            group_owner = metadata.owner
-        else:
-            group_owner = User.get_public_user()
-
-        if metadata.tus_transaction_id and not metadata.files:
-            raise HoustonException(
-                log,
-                'Tus transaction AssetGroup must contain files',
-            )
-        if not metadata.files and not group_owner.is_researcher:
-            raise HoustonException(
-                log,
-                'Only a Researcher can create an AssetGroup without any Assets',
-            )
-        asset_group = AssetGroup(
-            major_type=AssetGroupMajorType.filesystem,
-            description=metadata.description,
-            owner_guid=group_owner.guid,
-        )
-
-        if metadata.anonymous_submitter:
-            asset_group.submitter = metadata.anonymous_submitter
-
-        with db.session.begin(subtransactions=True):
-            db.session.add(asset_group)
-
-        log.debug('created asset_group %r' % asset_group)
-
-        if metadata.tus_transaction_id:
-            try:
-                added = asset_group.import_tus_files(
-                    transaction_id=metadata.tus_transaction_id, paths=metadata.files
-                )
-            except Exception:
-                log.exception(
-                    'create_from_tus() had problems with import_tus_files(); deleting from db and fs %r'
-                    % asset_group
-                )
-                asset_group.delete()
-                raise
-
-            log.debug('asset_group imported %r' % added)
-        return asset_group
-
-    @classmethod
-    def create_from_tus(
-        cls, description, owner, transaction_id, paths=None, submitter=None
-    ):
-        assert transaction_id is not None
-        if owner is not None and not owner.is_anonymous:
-            group_owner = owner
-        else:
-            group_owner = User.get_public_user()
-        asset_group = AssetGroup(
-            major_type=AssetGroupMajorType.filesystem,
-            description=description,
-            owner_guid=group_owner.guid,
-        )
-
-        if submitter:
-            asset_group.submitter = submitter
-
-        with db.session.begin(subtransactions=True):
-            db.session.add(asset_group)
-
-        log.info('created asset_group %r' % asset_group)
-        added = None
-        try:
-            added = asset_group.import_tus_files(
-                transaction_id=transaction_id, paths=paths
-            )
-        except Exception:
-            log.exception(
-                'create_from_tus() had problems with import_tus_files(); deleting from db and fs %r'
-                % asset_group
-            )
-            asset_group.delete()
-            raise
-
-        log.info('asset_group imported %r' % added)
-        return asset_group
-
-    def import_tus_files(self, transaction_id=None, paths=None, purge_dir=True):
-        from app.extensions.tus import _tus_filepaths_from, _tus_purge
-        from .tasks import git_push
-
-        self.ensure_repository()
-
-        sub_id = None if transaction_id is not None else self.guid
-        asset_group_abspath = self.get_absolute_path()
-        asset_group_path = os.path.join(asset_group_abspath, '_asset_group')
-        paths_added = []
-        num_files = 0
-
-        for path in _tus_filepaths_from(
-            asset_group_guid=sub_id, transaction_id=transaction_id, paths=paths
-        ):
-            name = pathlib.Path(path).name
-            paths_added.append(name)
-            num_files += 1
-            os.rename(path, os.path.join(asset_group_path, name))
-
-        assets_added = []
-        if num_files > 0:
-            log.debug('Tus collect for %d files moved' % (num_files))
-            self.git_commit('Tus collect commit for %d files.' % (num_files,))
-            # Do git push to gitlab in the background (we won't wait for its
-            # completion here)
-            git_push.delay(str(self.guid))
-            for asset in self.assets:
-                if asset.path in paths_added:
-                    assets_added.append(asset)
-
-        if purge_dir:
-            # may have some unclaimed files in it
-            _tus_purge(asset_group_guid=sub_id, transaction_id=transaction_id)
-        return assets_added
-
-    def realize_asset_group(self):
-        """
-        Unpack any archives and resolve any symlinks
-
-        Must check for security vulnerabilities around decompression bombs and
-        recursive links
-        """
-        ARCHIVE_MIME_TYPE_WHITELIST = [  # NOQA
-            'application/gzip',
-            'application/vnd.rar',
-            'application/x-7z-compressed',
-            'application/x-bzip',
-            'application/x-bzip2',
-            'application/x-tar',
-            'application/zip',
-        ]
-        pass
-
-    def update_asset_symlinks(self, verbose=True, existing_filepath_guid_mapping={}):
-        """
-        Traverse the files in the _asset_group/ folder and add/update symlinks
-        for any relevant files we identify
-
-        Ref:
-            https://pypi.org/project/python-magic/
-            https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Common_types
-            http://www.iana.org/assignments/media-types/media-types.xhtml
-        """
-        import utool as ut
-        import magic
-
-        asset_group_abspath = self.get_absolute_path()
-        asset_group_path = os.path.join(asset_group_abspath, '_asset_group')
-        assets_path = os.path.join(asset_group_abspath, '_assets')
-
-        # Walk the asset_group path, looking for white-listed MIME type files
-        files = []
-        skipped = []
-        errors = []
-        walk_list = sorted(list(os.walk(asset_group_path)))
-        for root, directories, filenames in tqdm.tqdm(walk_list):
-            filenames = sorted(filenames)
-            for filename in filenames:
-                filepath = os.path.join(root, filename)
-
-                # Normalize path (sanity check)
-                filepath = os.path.normpath(filepath)
-
-                # Sanity check, ensure that the path is formatted well
-                assert os.path.exists(filepath)
-                assert os.path.isabs(filepath)
-                try:
-                    basename = os.path.basename(filepath)
-                    _, extension = os.path.splitext(basename)
-                    extension = extension.lower()
-                    extension = extension.strip('.')
-
-                    if basename.startswith('.'):
-                        # Skip hidden files
-                        if basename not in ['.touch']:
-                            skipped.append((filepath, basename))
-                        continue
-
-                    if os.path.isdir(filepath):
-                        # Skip any directories (sanity check)
-                        skipped.append((filepath, extension))
-                        continue
-
-                    if os.path.islink(filepath):
-                        # Skip any symbolic links (sanity check)
-                        skipped.append((filepath, extension))
-                        continue
-                    mime_type = magic.from_file(filepath, mime=True)
-                    if mime_type not in self.mime_type_whitelist:
-                        # Skip any unsupported MIME types
-                        skipped.append((filepath, extension))
-                        continue
-
-                    magic_signature = magic.from_file(filepath)
-                    size_bytes = os.path.getsize(filepath)
-
-                    file_data = {
-                        'filepath': filepath,
-                        'path': basename,
-                        'extension': extension,
-                        'mime_type': mime_type,
-                        'magic_signature': magic_signature,
-                        'size_bytes': size_bytes,
-                        'asset_group_guid': self.guid,
-                    }
-                    files.append(file_data)
-                except Exception:
-                    logging.exception('Got exception in update_asset_symlinks')
-                    errors.append(filepath)
-
-        if verbose:
-            print('Processed asset files from asset_group: %r' % (self,))
-            print('\tFiles   : %d' % (len(files),))
-            print('\tSkipped : %d' % (len(skipped),))
-            if len(skipped) > 0:
-                skipped_ext_list = [skip[1] for skip in skipped]
-                skipped_ext_str = ut.repr3(ut.dict_hist(skipped_ext_list))
-                skipped_ext_str = skipped_ext_str.replace('\n', '\n\t\t')
-                print('\t\t%s' % (skipped_ext_str,))
-            print('\tErrors  : %d' % (len(errors),))
-
-        # Compute the xxHash64 for all found files
-        filepath_list = [file_data_['filepath'] for file_data_ in files]
-        arguments_list = list(zip(filepath_list))
-        print('Computing filesystem xxHash64...')
-        filesystem_xxhash64_list = parallel(
-            compute_xxhash64_digest_filepath, arguments_list
-        )
-        filesystem_guid_list = list(map(ut.hashable_to_uuid, filesystem_xxhash64_list))
-
-        # Update file_data with the filesystem and semantic hash information
-        zipped = zip(files, filesystem_xxhash64_list, filesystem_guid_list)
-        for file_data, filesystem_xxhash64, filesystem_guid in zipped:
-            file_data['filesystem_xxhash64'] = filesystem_xxhash64
-            file_data['filesystem_guid'] = filesystem_guid
-            semantic_guid_data = (
-                file_data['asset_group_guid'],
-                file_data['filesystem_guid'],
-            )
-            file_data['semantic_guid'] = ut.hashable_to_uuid(semantic_guid_data)
-
-        # Delete all existing symlinks
-        existing_asset_symlinks = ut.glob(os.path.join(assets_path, '*'))
-        for existing_asset_symlink in existing_asset_symlinks:
-            basename = os.path.basename(existing_asset_symlink)
-            if basename in ['.touch', 'derived']:
-                continue
-            existing_asset_target = os.readlink(existing_asset_symlink)
-            existing_asset_target_ = os.path.abspath(
-                os.path.join(assets_path, existing_asset_target)
-            )
-            if os.path.exists(existing_asset_target_):
-                uuid_str, _ = os.path.splitext(basename)
-                uuid_str = uuid_str.strip().strip('.')
-                if existing_asset_target_ not in existing_filepath_guid_mapping:
-                    try:
-                        existing_filepath_guid_mapping[
-                            existing_asset_target_
-                        ] = uuid.UUID(uuid_str)
-                    except Exception:
-                        pass
-            os.remove(existing_asset_symlink)
-
-        # Add new or update any existing Assets found in the AssetGroup
-        asset_asset_group_filepath_list = [
-            file_data.pop('filepath', None) for file_data in files
-        ]
-        assets = []
-        # TODO: slim down this DB context
-        with db.session.begin(subtransactions=True):
-            for file_data, asset_asset_group_filepath in zip(
-                files, asset_asset_group_filepath_list
-            ):
-                semantic_guid = file_data.get('semantic_guid', None)
-                asset = Asset.query.filter(Asset.semantic_guid == semantic_guid).first()
-                if asset is None:
-                    # Check if we can recycle existing GUID from symlink
-                    recycle_guid = existing_filepath_guid_mapping.get(
-                        asset_asset_group_filepath, None
-                    )
-                    if recycle_guid is not None:
-                        file_data['guid'] = recycle_guid
-                    # Create record if asset is new
-                    asset = Asset(**file_data)
-                    db.session.add(asset)
-                else:
-                    # Update record if Asset exists
-                    for key in file_data:
-                        if key in [
-                            'asset_group_guid',
-                            'filesystem_guid',
-                            'semantic_guid',
-                        ]:
-                            continue
-                        value = file_data[key]
-                        setattr(asset, key, value)
-                    db.session.merge(asset)
-                assets.append(asset)
-
-        # Update all symlinks for each Asset
-        for asset, asset_asset_group_filepath in zip(
-            assets, asset_asset_group_filepath_list
-        ):
-            db.session.refresh(asset)
-            asset.update_symlink(asset_asset_group_filepath)
-            asset.set_derived_meta()
-            if verbose:
-                print(filepath)
-                print('\tAsset         : %s' % (asset,))
-                print('\tSemantic GUID : %s' % (asset.semantic_guid,))
-                print('\tExtension     : %s' % (asset.extension,))
-                print('\tMIME type     : %s' % (asset.mime_type,))
-                print('\tSignature     : %s' % (asset.magic_signature,))
-                print('\tSize bytes    : %s' % (asset.size_bytes,))
-                print('\tFS xxHash64   : %s' % (asset.filesystem_xxhash64,))
-                print('\tFS GUID       : %s' % (asset.filesystem_guid,))
-
-        # Get all historical and current Assets for this AssetGroup
-        db.session.refresh(self)
-
-        # Delete any historical Assets that have been deleted from this commit
-        deleted_assets = list(set(self.assets) - set(assets))
-        if verbose:
-            print('Deleting %d orphaned Assets' % (len(deleted_assets),))
-        with db.session.begin(subtransactions=True):
-            for deleted_asset in deleted_assets:
-                deleted_asset.delete()
-        db.session.refresh(self)
-
-    def update_metadata_from_project(self, project):
-        # Update any local metadata from sub
-        for tag in project.tag_list:
-            tag = tag.strip().split(':')
-            if len(tag) == 2:
-                key, value = tag
-                key_ = key.lower()
-                value_ = value.lower()
-                if key_ == 'type':
-                    default_major_type = AssetGroupMajorType.unknown
-                    self.major_type = getattr(
-                        AssetGroupMajorType, value_, default_major_type
-                    )
-
-        self.description = project.description
-        with db.session.begin():
-            db.session.merge(self)
-        db.session.refresh(self)
-
-    def update_metadata_from_repo(self, repo):
-        repo = self.get_repository()
-        assert repo is not None
-
-        if len(repo.branches) > 0:
-            commit = repo.branches.master.commit
-            self.update_metadata_from_commit(commit)
-
-        return repo
-
-    def update_metadata_from_commit(self, commit):
-        self.commit = commit.hexsha
-
-        metadata_path = os.path.join(commit.repo.working_dir, 'metadata.json')
-        assert os.path.exists(metadata_path)
-        with open(metadata_path, 'r') as metadata_file:
-            metadata_dict = json.load(metadata_file)
-
-        self.commit_mime_whitelist_guid = metadata_dict.get(
-            'commit_mime_whitelist_guid', self.mime_type_whitelist_guid
-        )
-        self.commit_houston_api_version = metadata_dict.get(
-            'commit_houston_api_version', version
-        )
-
-        with db.session.begin(subtransactions=True):
-            db.session.merge(self)
-        db.session.refresh(self)
-
-    def get_absolute_path(self):
-        asset_group_database_path = current_app.config.get(
-            'ASSET_GROUP_DATABASE_PATH', None
-        )
-        assert asset_group_database_path is not None
-        assert os.path.exists(asset_group_database_path)
-
-        asset_group_path = os.path.join(asset_group_database_path, str(self.guid))
-
-        return asset_group_path
-
-    def delete_dirs(self):
-        if os.path.exists(self.get_absolute_path()):
-            shutil.rmtree(self.get_absolute_path())
+            local_store_metadata['frontend_sightings_data'] = metadata_request
 
     def is_partially_in_stage(self, stage):
         if self.asset_group_sightings:
@@ -1582,12 +829,6 @@ class AssetGroup(db.Model, HoustonModel):
 
     def is_processed(self):
         return self.is_completely_in_stage(AssetGroupSightingStage.processed)
-
-    def get_asset_for_file(self, filename):
-        for asset in self.assets:
-            if asset.path == filename:
-                return asset
-        return None
 
     def get_asset_group_sightings_for_asset(self, asset):
         return [ags for ags in self.asset_group_sightings if ags.has_filename(asset.path)]
@@ -1637,21 +878,10 @@ class AssetGroup(db.Model, HoustonModel):
             ags.asset_updated(asset)
 
     def delete(self):
-        from .tasks import delete_remote
-
         with db.session.begin(subtransactions=True):
-            for asset in self.assets:
-                asset.delete_cascade()
             for sighting in self.asset_group_sightings:
                 sighting.delete()
-        # TODO: This is potentially dangerous as it decouples the Asset deletion
-        #       transaction with the AssetGroup deletion transaction, bad for rollbacks
-        with db.session.begin(subtransactions=True):
-            db.session.delete(self)
-        self.delete_dirs()
-        # Delete the gitlab project in the background (we won't wait
-        # for its completion)
-        delete_remote.delay(str(self.guid))
+        super(AssetGroup, self).delete()
 
     def delete_asset_group_sighting(self, asset_group_sighting):
         with db.session.begin(subtransactions=True):
@@ -1660,52 +890,3 @@ class AssetGroup(db.Model, HoustonModel):
                 if asset_ags == [asset_group_sighting]:
                     asset.delete_cascade()
             asset_group_sighting.delete()
-
-    # stub of DEX-220 ... to be continued
-    def justify_existence(self):
-        if self.assets:  # we have assets, so we live on
-            return
-        log.warning('justify_existence() found ZERO assets, self-destructing %r' % self)
-        self.delete()  # TODO will this also kill remote repo?
-
-    def get_repository(self):
-        repo_path = pathlib.Path(self.get_absolute_path())
-        if (repo_path / '.git').exists():
-            return Repo(repo_path)
-
-    def ensure_repository(self):
-        from app.extensions.elapsed_time import ElapsedTime
-
-        timer = ElapsedTime()
-        repo = self.get_repository()
-        if repo:
-            if 'origin' in repo.remotes:
-                repo = self.git_pull()
-        else:
-            project = AssetGroup.get_remote(self.guid)
-            if project:
-                repo = self.git_clone(project)
-            else:
-                repo = Repo.init(self.get_absolute_path())
-        self._ensure_repository_files()
-
-        if float(timer.elapsed()) > 0.05:
-            log.info(
-                f'Ensure Git repository for AssetGroup {self.guid} took {timer.elapsed()} seconds'
-            )
-        return repo
-
-    @classmethod
-    def get_remote(cls, guid):
-        try:
-            return current_app.git_backend.get_project(str(guid))
-        except (GitlabInitializationError, requests.exceptions.RequestException):
-            log.exception(f'Error when calling AssetGroup.get_remote({guid})')
-
-    @classmethod
-    def is_on_remote(cls, guid):
-        try:
-            return current_app.git_backend.is_project_on_remote(str(guid))
-        except (GitlabInitializationError, requests.exceptions.RequestException):
-            log.exception(f'Error when calling AssetGroup.is_on_remote({guid})')
-            return False
