+++ app/modules/asset_groups/schemas.py
@@ -6,7 +6,12 @@ Serialization schemas for Asset_groups resources RESTful API
 
 from flask_marshmallow import base_fields
 from flask_restx_patched import ModelSchema
-from app.modules.assets.schemas import ExtendedAssetDetailedAnnotationsSchema
+from app.modules.assets.schemas import (
+    ExtendedAssetDetailedAnnotationsSchema,
+    BaseGitStoreSchema,
+    CreateGitStoreSchema,
+    DetailedGitStoreSchema,
+)
 
 from .models import AssetGroup, AssetGroupSighting
 
@@ -27,6 +32,37 @@ SIGHTING_FIELDS_IN_AGS_CONFIG = {
     'comments',
     'featuredAssetGuid',
 }
+
+
+class BaseAssetGroupSchema(BaseGitStoreSchema):
+    """
+    Base Asset_group schema exposes only the most general fields.
+    """
+
+    pass
+
+
+class CreateAssetGroupSchema(CreateGitStoreSchema):
+    """
+    Detailed Asset_group schema exposes all useful fields.
+    """
+
+    pass
+
+
+class DetailedAssetGroupSchema(DetailedGitStoreSchema):
+    """
+    Detailed Asset_group schema exposes all useful fields.
+    """
+
+    asset_group_sightings = base_fields.Nested(
+        'DetailedAssetGroupSightingSchema',
+        many=True,
+    )
+
+    class Meta(DetailedGitStoreSchema.Meta):
+        fields = DetailedGitStoreSchema.Meta.fields + ('asset_group_sightings',)
+        dump_only = DetailedGitStoreSchema.Meta.dump_only
 
 
 class BaseAssetGroupSightingSchema(ModelSchema):
@@ -200,66 +236,6 @@ class AssetGroupSightingAsSightingSchema(ModelSchema):
             AssetGroupSighting.asset_group_guid.key,
         )
         dump_only = BaseAssetGroupSightingSchema.Meta.dump_only
-
-
-class BaseAssetGroupSchema(ModelSchema):
-    """
-    Base Asset_group schema exposes only the most general fields.
-    """
-
-    class Meta:
-        # pylint: disable=missing-docstring
-        model = AssetGroup
-        fields = (
-            AssetGroup.guid.key,
-            AssetGroup.commit.key,
-            AssetGroup.major_type.key,
-            AssetGroup.description.key,
-        )
-        dump_only = (
-            AssetGroup.guid.key,
-            AssetGroup.commit.key,
-        )
-
-
-class CreateAssetGroupSchema(BaseAssetGroupSchema):
-    """
-    Detailed Asset_group schema exposes all useful fields.
-    """
-
-    class Meta(BaseAssetGroupSchema.Meta):
-        fields = BaseAssetGroupSchema.Meta.fields + (
-            AssetGroup.owner_guid.key,
-            AssetGroup.created.key,
-            AssetGroup.updated.key,
-        )
-        dump_only = BaseAssetGroupSchema.Meta.dump_only + (
-            AssetGroup.owner_guid.key,
-            AssetGroup.created.key,
-            AssetGroup.updated.key,
-        )
-
-
-class DetailedAssetGroupSchema(CreateAssetGroupSchema):
-    """
-    Detailed Asset_group schema exposes all useful fields.
-    """
-
-    from app.modules.assets.models import Asset
-
-    assets = base_fields.Nested(
-        'BaseAssetSchema',
-        exclude=Asset.asset_group_guid.key,
-        many=True,
-    )
-    asset_group_sightings = base_fields.Nested(
-        'DetailedAssetGroupSightingSchema',
-        many=True,
-    )
-
-    class Meta(CreateAssetGroupSchema.Meta):
-        fields = CreateAssetGroupSchema.Meta.fields + ('assets', 'asset_group_sightings')
-        dump_only = CreateAssetGroupSchema.Meta.dump_only
 
 
 class DebugAssetGroupSightingSchema(DetailedAssetGroupSightingSchema):
