+++ app/modules/assets/resources.py
@@ -146,9 +146,9 @@ class AssetSrcUByID(Resource):
         },
     )
     def get(self, asset, format):
-        from app.modules.asset_groups.models import AssetGroup
+        cls = type(asset.git_store)
+        cls.ensure_store(asset.git_store_guid)
 
-        AssetGroup.ensure_asset_group(asset.asset_group_guid)
         try:
             asset_format_path = asset.get_or_make_format_path(format)
         except Exception:
