+++ app/modules/sightings/models.py
@@ -463,6 +463,7 @@ class Sighting(db.Model, FeatherModel):
     def get_augmented_sighting_json(self):
         if is_extension_enabled('edm'):
             response = current_app.edm.get_dict('sighting.data_complete', self.guid)
+
         if not isinstance(response, dict):  # some non-200 thing, incl 404
             return response
         if not response.get('success', False):
